🔄 refactor: migrate apps/web from Next.js to Docusaurus
- replace the Next.js landing page with a Docusaurus site, porting the existing marketing homepage (hero, feature grid, component categories) into src/pages/index.tsx
- wire ui-kit's own Tailwind v4 build into Docusaurus via a configurePostCss plugin, since @repo/ui/style.css resolves to source CSS (not pre-built utilities) for workspace consumers
- bridge Docusaurus' color mode into ui-kit's Config via a DemoTheme wrapper (same pattern used in use-hooks/live-editor) so the navbar's dark/light toggle drives the ui-kit components on the page
- drop the app's "type": "module" package.json field — combined with Docusaurus' webpack SSR bundle, it caused a require.resolveWeak is not a function build failure
- add apps/web/vercel.json (framework: null, custom build/output) matching apps/docs' deploy pattern, since Vercel no longer auto-detects Next.js here
- update turbo.json to cache Docusaurus' build/ output, and refresh README/README.ko/CLAUDE.md references to apps/web

apps/docs (Storybook) is untouched.
