+++ apps/web/.gitignore
@@ -1,36 +1,9 @@
-# See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
+# Generated files
+.docusaurus
+.cache-loader
 
-# dependencies
-/node_modules
-/.pnp
-.pnp.js
-.yarn/install-state.gz
-
-# testing
-/coverage
-
-# next.js
-/.next/
-/out/
-
-# production
-/build
-
-# misc
-.DS_Store
-*.pem
-
-# debug
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
-
-# env files (can opt-in for commiting if needed)
-.env*
-
-# vercel
-.vercel
-
-# typescript
-*.tsbuildinfo
-next-env.d.ts
+# Misc
+.env.local
+.env.development.local
+.env.test.local
+.env.production.local
