--- apps/web/app/globals.css
@@ -1,25 +1,0 @@
-@import 'tailwindcss';
-@import '@repo/ui/style.css';
-
-html {
-  min-height: 100%;
-  color-scheme: light;
-}
-
-html.dark {
-  color-scheme: dark;
-}
-
-body {
-  /* Theme body colors used to come from `@repo/ui/style.css`; that global
-     `body {}` rule was removed from the library (a component lib shouldn't
-     paint the host body), so the app sets them itself now. */
-  @apply bg-background text-foreground;
-  max-width: 100vw;
-  overflow-x: hidden;
-}
-
-a {
-  color: inherit;
-  text-decoration: none;
-}
