--- apps/web/app/layout.tsx
@@ -1,50 +1,0 @@
-import type { Metadata } from 'next';
-import localFont from 'next/font/local';
-
-import './globals.css';
-
-const geistSans = localFont({
-  src: './fonts/GeistVF.woff',
-  variable: '--font-geist-sans',
-});
-const geistMono = localFont({
-  src: './fonts/GeistMonoVF.woff',
-  variable: '--font-geist-mono',
-});
-
-export const metadata: Metadata = {
-  title: 'ui-kit — React UI component library',
-  description:
-    'Modern React UI component library built with TypeScript, Tailwind CSS, and Radix UI. Featuring atoms, molecules, organisms and layout templates for building beautiful interfaces.',
-};
-
-const themeInitScript = `
-(function () {
-  try {
-    var stored = localStorage.getItem('ui-kit-web-theme');
-    var theme = stored ? JSON.parse(stored) : 'dark';
-    if (theme === 'dark') {
-      document.documentElement.classList.add('dark');
-    }
-  } catch (e) {
-    document.documentElement.classList.add('dark');
-  }
-})();
-`;
-
-export default function RootLayout({
-  children,
-}: Readonly<{
-  children: React.ReactNode;
-}>) {
-  return (
-    <html lang="en">
-      <head>
-        <script dangerouslySetInnerHTML={{ __html: themeInitScript }} />
-      </head>
-      <body className={`${geistSans.variable} ${geistMono.variable}`}>
-        {children}
-      </body>
-    </html>
-  );
-}
