+++ apps/web/eslint.config.mjs
@@ -1,0 +1,9 @@
+import { config } from '@repo/eslint-config/react-internal';
+
+/** @type {import("eslint").Linter.Config} */
+export default [
+  ...config,
+  {
+    ignores: ['.docusaurus/**', 'build/**'],
+  },
+];
