+++ apps/web/src/css/custom.css
@@ -1,0 +1,20 @@
+/**
+ * Any CSS included here will be global. The classic template
+ * bundles Infima by default. Infima is a CSS framework designed to
+ * work well for content-centric websites.
+ */
+
+/*
+ * `@repo/ui/style.css` resolves to the library's source globals.css inside
+ * this workspace (theme tokens + utilities layer, no preflight — see that
+ * file for why), so importing `tailwindcss` here is what actually turns
+ * this page's own utility classes (`flex`, `max-w-5xl`, `grid-cols-3`, ...)
+ * into real CSS; the `configurePostCss` hook in docusaurus.config.ts is
+ * what runs the Tailwind build itself.
+ */
+@import 'tailwindcss';
+@import '@repo/ui/style.css';
+
+body {
+  @apply bg-background text-foreground;
+}
