+++ apps/web/vercel.json
@@ -1,0 +1,5 @@
+{
+  "framework": null,
+  "buildCommand": "pnpm build",
+  "outputDirectory": "build"
+}
